chore: added student grade report entry overlay to grades report details
- moved overlay related code from `ReportCardLive.StudentsGradesComponent` to a shared component `GradesReports.StudentGradeReportEntryOverlayComponent`
